pyproject.toml
[tool.poetry]
name = "gredditwallpaper"
version = "0.1.0"
description = "Download and set a desktop background in gnome using a random image from a chosen subreddit via CLI or GUI."
authors = ["AJ Bucci <ajbucci@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
PyGObject = "^3.46.0"
requests = "^2.31.0"
Pillow = "^10.1.0"
isort = "^5.13.2"
types-requests = "^2.31.0.10"

[tool.poetry.group.test.dependencies]
pytest = "^7.4.3"

[tool.poetry.group.lint.dependencies]
black = "^23.10.1"
flake8-bugbear = "23.3.12" # over this version python 3.8.1 is required
mypy = "^1.6.1"
pep8-naming = "^0.13.3"
vulture = "^2.10"
bandit = "^1.7.5"
safety = "^2.3.5"
isort = "^5.13.2"

[tool.poetry.scripts]
gredwall = "gredditwallpaper.gui:start"
gredwallc = "gredditwallpaper.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 150

[tool.isort]
profile = "hug"